+++ packages/blocks/core/CHANGELOG.md
@@ -41,6 +41,7 @@
 - `Post Content` block: inner blocks improved.
 - `Post Date` block: inner blocks improved.
 - `Post Excerpt` block: inner blocks improved.
+- `Post Terms` block: inner blocks improved.
 
 
 ### Automated Tests